Move activePeerSet and activePeerReset to actions
